Refactored game controller + service.
