+++ src/main/java/org/wlcp/wlcpapi/service/GameService.java
@@ -1,0 +1,25 @@
+package org.wlcp.wlcpapi.service;
+
+import java.util.List;
+
+import org.wlcp.wlcpapi.datamodel.master.Game;
+import org.wlcp.wlcpapi.dto.CopyRenameDeleteGameDto;
+import org.wlcp.wlcpapi.dto.GameDto;
+
+public interface GameService {
+	
+	public List<GameDto> getPrivateGames(String usernameId);
+	
+	public List<GameDto> getPublicGames();
+	
+	public Game loadGame(String gameId);
+	
+	public Game saveGame(Game game);
+	
+	public String copyGame(CopyRenameDeleteGameDto copyRenameDeleteGameDto);
+	
+	public String renameGame(CopyRenameDeleteGameDto copyRenameDeleteGameDto);
+	
+	public String deleteGame(CopyRenameDeleteGameDto copyRenameDeleteGameDto);
+
+}
