--- src/main/java/org/wlcp/wlcpapi/service/SaveGameService.java
@@ -1,9 +1,0 @@
-package org.wlcp.wlcpapi.service;
-
-import org.wlcp.wlcpapi.datamodel.master.Game;
-
-public interface SaveGameService {
-	
-	public Game saveGame(Game game);
-
-}
